Add tests/jit folder, split jit in models/filterbanks
